revise: moves the existing backend configuration to their own modules
