+++ wdl-engine/src/backend/docker.rs
@@ -48,8 +48,8 @@ use crate::STDOUT_FILE_NAME;
 use crate::Value;
 use crate::WORK_DIR_NAME;
 use crate::config::DEFAULT_TASK_SHELL;
-use crate::config::DockerBackendConfig;
 use crate::config::TaskConfig;
+use crate::config::backend::DockerBackendConfig;
 use crate::http::Downloader;
 use crate::http::HttpDownloader;
 use crate::http::Location;
