+++ wdl-engine/src/backend/local.rs
@@ -36,8 +36,8 @@ use crate::TaskExecutionResult;
 use crate::Value;
 use crate::WORK_DIR_NAME;
 use crate::config::DEFAULT_TASK_SHELL;
-use crate::config::LocalBackendConfig;
 use crate::config::TaskConfig;
+use crate::config::backend::LocalBackendConfig;
 use crate::convert_unit_string;
 use crate::http::Downloader;
 use crate::http::HttpDownloader;
