+++ wdl-engine/src/config/backend/local.rs
@@ -1,0 +1,76 @@
+//! Configuration for the local backend.
+
+use anyhow::Context as _;
+use anyhow::Result;
+use anyhow::bail;
+use serde::Deserialize;
+use serde::Serialize;
+
+use crate::SYSTEM;
+use crate::convert_unit_string;
+
+/// Represents configuration for the local task execution backend.
+///
+/// <div class="warning">
+/// Warning: the local task execution backend spawns processes on the host
+/// directly without the use of a container; only use this backend on trusted
+/// WDL. </div>
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case", deny_unknown_fields)]
+pub struct LocalBackendConfig {
+    /// Set the number of CPUs available for task execution.
+    ///
+    /// Defaults to the number of logical CPUs for the host.
+    ///
+    /// The value cannot be zero or exceed the host's number of CPUs.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub cpu: Option<u64>,
+
+    /// Set the total amount of memory for task execution as a unit string (e.g.
+    /// `2 GiB`).
+    ///
+    /// Defaults to the total amount of memory for the host.
+    ///
+    /// The value cannot be zero or exceed the host's total amount of memory.
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub memory: Option<String>,
+}
+
+impl LocalBackendConfig {
+    /// Validates the local task execution backend configuration.
+    pub fn validate(&self) -> Result<()> {
+        if let Some(cpu) = self.cpu {
+            if cpu == 0 {
+                bail!("local backend configuration value `cpu` cannot be zero");
+            }
+
+            let total = SYSTEM.cpus().len() as u64;
+            if cpu > total {
+                bail!(
+                    "local backend configuration value `cpu` cannot exceed the virtual CPUs \
+                     available to the host ({total})"
+                );
+            }
+        }
+
+        if let Some(memory) = &self.memory {
+            let memory = convert_unit_string(memory).with_context(|| {
+                format!("local backend configuration value `memory` has invalid value `{memory}`")
+            })?;
+
+            if memory == 0 {
+                bail!("local backend configuration value `memory` cannot be zero");
+            }
+
+            let total = SYSTEM.total_memory();
+            if memory > total {
+                bail!(
+                    "local backend configuration value `memory` cannot exceed the total memory of \
+                     the host ({total} bytes)"
+                );
+            }
+        }
+
+        Ok(())
+    }
+}
